workign on bunny

+++ Source/StopMotion/GraphCutMethods/GCOLib/GCoptimization.h
@@ -167,7 +167,7 @@ public:
 #else
 	typedef double EnergyType;  // 64-bit energy total
 #endif
-	typedef float EnergyTermType;    // 32-bit energy terms
+	typedef double EnergyTermType;    // 32-bit energy terms
 #endif
 	typedef Energy<EnergyTermType,EnergyTermType,EnergyType> EnergyT;
 	typedef EnergyT::Var VarID;
